app/requirements.txt
streamlit>=1.28
pandas>=2.0
numpy>=1.26
plotly>=6.0
scikit-learn>=1.3
psycopg2-binary>=2.9
python-dotenv>=1.0
sqlalchemy